Show tips instead of random loading messages
