--- plugins/woo-ai/src/components/index.ts
@@ -1,3 +1,0 @@
-export * from './error-message';
-export * from './random-loading-message';
-export * from './random-tip-message';
